Chassis Chain support for IxNetwork only (#592)
* Initial submit for chassischain suport in snappi-IxNetwork.
Includes the class required for chassishcain within snappi-ixnetwork.

* Adding the converserion logic.
Once the chassis are added and if they are in chassis chain configuration and the same is specified in the test the parameters will be set according to the configuration provided.

* Adding basic IPv4 testcases with chassis chain support.
One pair of chassis chain configuration is present and consecutive chassis chain if required are commented out for reference

* Using the chassis_chain directly from from the api. Also clearing the chassis_chain as it is not being cleared after the test runs.

* This commit contains the below:-
1. Renamed the class.
2. Added docstring
3. Adding validation in the test and marking it to skip in the CI.